+++ src/services/storage.ts
@@ -6,8 +6,8 @@ const dir = path.dirname(require.main.filename);
 export class Storage {
     public static get(...filePath: string[]): Promise<any> {
         return new Promise((resolve, reject) => {
-            const fileName = path.join(dir, ...filePath);
-            fs.readFile(path.resolve(__dirname, fileName), 'utf8', (error, data) => {
+            const fileName = this.makePath(...filePath);
+            fs.readFile(fileName, 'utf8', (error, data) => {
                 if (error) {
                     reject(error);
                     return;
@@ -17,10 +17,14 @@ export class Storage {
         });
     }
 
+    public static makePath(...filePath: string[]): string {
+        return path.resolve(__dirname, path.join(dir, ...filePath));
+    }
+
     public static set(data: object, ...filePath: string[]): Promise<any> {
         return new Promise((resolve, reject) => {
-            const fileName = path.join(dir, ...filePath);
-            fs.writeFile(path.resolve(__dirname, fileName), JSON.stringify(data), (error) => {
+            const fileName = this.makePath(...filePath);
+            fs.writeFile(fileName, JSON.stringify(data), (error) => {
                 if (error) {
                     reject(error);
                 } else {